requirements.txt
streamlit
numpy
opencv-python-headless
face-recognition
pillow